Modif makefile pour y inclure la libft

+++ ft_printf_precision.c
@@ -6,7 +6,7 @@
 /*   By: tomlulu <tomlulu@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/14 10:08:03 by tomlulu           #+#    #+#             */
-/*   Updated: 2018/01/16 09:20:10 by tmaraval         ###   ########.fr       */
+/*   Updated: 2018/01/16 10:07:22 by tmaraval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ void	ft_printf_precision(t_parsed_opt *opt)
 			temp = opt->str_arg;
 			opt->str_arg = ft_strsub(opt->str_arg, 1, ft_strlen(opt->str_arg));
 			free(temp);
-			while (ft_strlen(opt->str_arg) < opt->in_precision)
+			while (ft_strlen(opt->str_arg) < (size_t)opt->in_precision)
 			{
 				temp = opt->str_arg;
 				opt->str_arg = ft_strjoin("0", opt->str_arg);
@@ -41,7 +41,7 @@ void	ft_printf_precision(t_parsed_opt *opt)
 		}
 		else
 		{
-			while (ft_strlen(opt->str_arg) < opt->in_precision)
+			while (ft_strlen(opt->str_arg) < (size_t)opt->in_precision)
 			{
 				temp = opt->str_arg;
 				opt->str_arg = ft_strjoin("0", opt->str_arg);
